Q2_hienpmhe180216/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
# If you plan to use pyodbc/ODBC driver for SQL Server
pyodbc==5.2.0
Flask-SQLAlchemy==3.0.3
# Optional: for colored terminal output on Windows
colorama>=0.4.4
